added the zircon computation routines to GeoParams & purged the previous routines
