+++ Project.toml
@@ -7,10 +7,14 @@ version = "0.2.8"
 BibTeX = "7b0aa2c9-049f-5cec-894a-2b6b781bb25e"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
+KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+Loess = "4345ca2d-374a-55d4-8d30-97f9976e7612"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
@@ -18,11 +22,14 @@ Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 BibTeX = "0.1.0"
 Interpolations = "0.13"
 LaTeXStrings = "1.2.1"
-Parameters = "0.12.0"
+Parameters = "0.12"
 Requires = "1.0 - 1.3"
 Setfield = "0.8 - 0.8"
 UnPack = "1.0"
 Unitful = "1.8 - 1.12"
+Loess = "0.5"
+KernelDensity= "0.6"
+StatsBase = "0.33"
 julia = "1.6"
 
 [extras]
